Added JFileChooser for choosing save path of log file
When logging is turned on, a JFileChooser window opens to specify the desired save path rather than always logging to a hard-coded file path. If logging is toggled off and then on again a new path can be specified. Error checking isn't great, but as long as you don't try to break it everything should work fine.